introduce NTrello::Client
This will wrap and eventually replace the Trello gem.
