+++ app/api_wrappers/n_trello.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+module NTrello
+end
