+++ spec/support/fake_api/trello/implementation.rb
@@ -3,10 +3,6 @@
 require_relative "../modules"
 
 class FakeApi::Trello::Implementation
-  def self.authorize_url(return_url:, **_args)
-    "/trello/tokens/new?returnUrl=#{return_url}"
-  end
-
   class Board
     def self.from_response(_response)
       all
