+++ spec/support/fake_apis_rails.rb
@@ -3,11 +3,13 @@
 if FakeApis.enabled?
   require_relative "fake_api/modules"
   require_relative "fake_api/github/implementation"
+  require_relative "fake_api/trello/client"
   require_relative "fake_api/trello/implementation"
 
   RSpec.configuration.before(js: true) do
     Integration::Github.implementation = FakeApi::Github::Implementation
     Integration::Trello.implementation = FakeApi::Trello::Implementation
+    Integration::Trello.client_class = FakeApi::Trello::Client
   end
 else
   # For integrations like GitHub we need a URL with a constant server port to
